Add fix createRecord -> numberSequenceUtilityies -> createRecord cycle
