+++ src/database/utilities/index.js
@@ -3,14 +3,18 @@
  * Sustainable Solutions (NZ) Ltd. 2019
  */
 
-export { createRecord } from './createRecord';
-export { deleteRecord } from './deleteRecord';
-export { mergeRecords } from './mergeRecords';
 export {
+  createRecord,
   getNextNumber,
   getNumberSequence,
-  NUMBER_SEQUENCE_KEYS,
   reuseNumber,
-} from './numberSequenceUtilities';
+  NUMBER_SEQUENCE_KEYS,
+} from './createRecord';
+export { deleteRecord } from './deleteRecord';
+// eslint-disable-next-line import/no-cycle
+export { mergeRecords } from './mergeRecords';
+
+// eslint-disable-next-line import/no-cycle
 export { getTotal, addBatchToParent, millisecondsToDays } from './utilities';
+
 export const MILLISECONDS_PER_DAY = 24 * 60 * 60 * 1000;
